+++ R/globals.R
@@ -213,17 +213,16 @@ xfrAnl <- function(fleOrg = "", fleTgt = "") {
     lstOrg <- zip::zip_list(fleOrg)$filename
     lst2Cp <- lstOrg[grepl("index.html|[0-9].*\\s[a-z].*?/", lstOrg)]
     lstCmb <- union(zip::zip_list(fleTgt)$filename, lst2Cp)
-    tmpDir <- tempdir()
+    xfrDir <- file.path(tempdir(), "xfrAnl")
 
     # create a list of files to be copied, extract them from the input file and
     # append them to the output file
-    zip::unzip(fleTgt,                 exdir = tmpDir)
-    zip::unzip(fleOrg, files = lst2Cp, exdir = tmpDir, overwrite = TRUE)
-    zip::zip(fleTgt,   files = lstCmb, root  = tmpDir)
+    zip::unzip(fleTgt,                 exdir = xfrDir)
+    zip::unzip(fleOrg, files = lst2Cp, exdir = xfrDir, overwrite = TRUE)
+    zip::zip(fleTgt,   files = lstCmb, root  = xfrDir)
 
     # remove the files and directories from the list of files to be copied
-    unlink(setdiff(unique(dirname(file.path(tmpDir, lstCmb))), tmpDir), recursive = TRUE)
-    unlink(file.path(tmpDir, lstCmb[basename(lstCmb) == lstCmb]))
+    unlink(xfrDir, recursive = TRUE)
 
     TRUE
 }