+++ tests/testthat/test-arrange_cols_omv.R
@@ -80,7 +80,7 @@ test_that("arrange_cols_omv works", {
     expect_true(chkFle(nmeOut, fleCnt = "meta"))
     expect_true(chkFle(nmeOut, fleCnt = "metadata.json"))
     expect_true(chkFle(nmeOut, fleCnt = "data.bin"))
-    dtaFrm <- read_omv(nmeOut, sveAtt = FALSE)
+    dtaFrm <- read_omv(nmeOut, sveAtt = FALSE, getSyn = TRUE)
     expect_s3_class(dtaFrm, "data.frame")
     expect_equal(dim(dtaFrm), c(60, 13))
     expect_equal(names(dtaFrm), c("Filter 1", "ID", "len", "logLen", "supp", "supp - Transform 1", "dose", "dose2", "Trial", "Residuals", "J", "K", "L"))
@@ -89,6 +89,10 @@ test_that("arrange_cols_omv works", {
     expect_equal(zip::zip_list(nmeOut)$filename,
       c("data.bin", "strings.bin", "meta", "metadata.json", "xdata.json", "index.html", "01 empty/analysis", "02 anova/analysis", "03 empty/analysis",
         "04 ancova/analysis", "05 empty/analysis", "02 anova/resources/3b518ea3d44f095f.png", "02 anova/resources/07288f96c58ae68b.png"))
+    expect_equal(attr(dtaFrm, "syntax"),
+      list(paste("jmv::ANOVA(formula = len ~ supp + dose2 + supp:dose2, data = data, effectSize = \"partEta\", modelTest = TRUE, qq = TRUE,",
+                 "contrasts = list(list(var=\"supp\", type=\"none\"), list(var=\"dose2\", type=\"polynomial\")), postHoc = ~ supp + dose2, emMeans = ~ dose2:supp)"),
+           "jmv::ancova(formula = len ~ supp + dose, data = data, effectSize = \"partEta\", modelTest = TRUE)"))
 
     unlink(nmeOut)
     # do not unlink nmeInp, this isn't a generated file, but a link