+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## Enhancements and bug fixes:
 * added `arrange_cols_omv` (to change the order of variables / columns in jamovi-files)
-* added the parameter `psvAnl` to preserve the analyses in data files (only for `arrange_cols_omv`, `merge_cols_omv`, `sort_omv` where it makes sense – there the number of rows / participants / units doesn't change)
+* added the parameter `psvAnl` to preserve the analyses in data files
+  (only for `arrange_cols_omv`, `merge_cols_omv`, `sort_omv` where it makes sense – there the number of rows / participants / units doesn't change)
 * bug fixes: corrected an error in `chkFle`, removed duplicate functions from `globals.R`